Added: Echo (Go) server-side adapter
Ahoi

First of all thanks for this awesome project. I really like the idea as a different approach to nowadays app development. Therefore, I wanted to contribute ^^

As I am a big friend of Svelte as a "compiled" client-side framework I thought it would be interesting to have Inertia running with a compiled server-side language and framework such as Echo (https://echo.labstack.com/). So, I went ahead and created a copy of the nice PingCRM-Svelte Demo by https://github.com/zgabievi/pingcrm-svelte and connected it with an Echo server. The middleware, readme and demo can all be found at https://github.com/elipZis/inertia-echo.

I do not know if this is helpful or not. For me, it was a personal project and grew to even more with vested interest to merge two fast & compiled frameworks.

Would be happy if just one person finds this helpful or interesting and am looking forward for input and maybe grow it even further :)


Cheers