Improve overflow behaviour
There is a full-width scroll bar visible on screens wide enough for the sponsors section. Setting the overflow behaviour to auto fixes this. The scrollbar is only visible if the content exceeds the available space now.

I saw this happening a couple of times. I guess the developer was on a Mac while writing the code? The scroll bars on Macs are similar to mobile scroll bars: Invisible while not scrolling. On Windows and Linux they are always visible. I try to attach a screenshot.